Django Ledger Settings Model Base Abstracts
